Decoupled augmentation base [partly finished]
