+++ kornia/augmentation/augmentation3d.py
@@ -6,26 +6,16 @@ from torch.nn.functional import pad
 
 from kornia.constants import Resample, BorderType
 from kornia.augmentation.augmentation import AugmentationBase
-from kornia.augmentation import functional3d as F
+from kornia.augmentation import functional as F
 from kornia.augmentation import random_generator as rg
+from kornia.augmentation.random_generator import AugParamDict
 from kornia.augmentation.utils import (
     _infer_batch_shape3d,
     _transform_input3d,
+    _validate_input_dtype,
     _tuple_range_reader,
     _singular_range_check
 )
-
-
-class AugmentationBase3D(AugmentationBase):
-    def __init__(self, p: float, return_transform: bool = False, same_one_batch: bool = False) -> None:
-        super(AugmentationBase3D, self).__init__(
-            p=p, return_transform=return_transform, same_on_batch=same_on_batch)
-
-    def infer_batch_shape(self, input: Union[torch.Tensor, Tuple[torch.Tensor, torch.Tensor]]) -> torch.Size:
-        return _infer_batch_shape3d(input)
-
-    def transform_tensor(self, input: torch.Tensor) -> torch.Tensor:
-        return _transform_input3d(input)
 
 
 class RandomHorizontalFlip3D(AugmentationBase3D):
@@ -66,16 +56,16 @@ class RandomHorizontalFlip3D(AugmentationBase3D):
     """
 
     def __init__(self, p: float = 0.5, return_transform: bool = False, same_on_batch: bool = False) -> None:
-        super(RandomHorizontalFlip3D, self).__init__(p, return_transform=return_transform, same_on_batch=same_on_batch)
+        super(RandomHorizontalFlip3D, self).__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch)
 
-    def generate_parameters(self, batch_shape: torch.Size) -> Dict[str, torch.Tensor]:
-        return rg.random_prob_generator(batch_shape[0], self.p, self.same_on_batch)
+    def generate_parameters(self, batch_shape: torch.Size) -> AugParamDict:
+        return AugParamDict(dict(params={}, flags={}))
 
-    def compute_transformation(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.compute_hflip_transformation3d(input, params)
+    def compute_transformation(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.compute_hflip_transformation3d(input)
 
-    def apply_transform(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.apply_hflip3d(input, params)
+    def apply_transform(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.apply_hflip3d(input)
 
 
 class RandomVerticalFlip3D(AugmentationBase3D):
@@ -115,16 +105,16 @@ class RandomVerticalFlip3D(AugmentationBase3D):
     """
 
     def __init__(self, p: float = 0.5, return_transform: bool = False, same_on_batch: bool = False) -> None:
-        super(RandomVerticalFlip3D, self).__init__(p, return_transform=return_transform, same_on_batch=same_on_batch)
+        super(RandomVerticalFlip3D, self).__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch)
 
-    def generate_parameters(self, batch_shape: torch.Size) -> Dict[str, torch.Tensor]:
-        return rg.random_prob_generator(batch_shape[0], self.p, self.same_on_batch)
+    def generate_parameters(self, batch_shape: torch.Size) -> AugParamDict:
+        return AugParamDict(dict(params={}, flags={}))
 
-    def compute_transformation(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.compute_vflip_transformation3d(input, params)
+    def compute_transformation(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.compute_vflip_transformation3d(input)
 
-    def apply_transform(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.apply_vflip3d(input, params)
+    def apply_transform(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.apply_vflip3d(input)
 
 
 class RandomDepthicalFlip3D(AugmentationBase3D):
@@ -165,13 +155,16 @@ class RandomDepthicalFlip3D(AugmentationBase3D):
     """
 
     def __init__(self, p: float = 0.5, return_transform: bool = False, same_on_batch: bool = False) -> None:
-        super(RandomDepthicalFlip3D, self).__init__(p, return_transform=return_transform, same_on_batch=same_on_batch)
+        super(RandomDepthicalFlip3D, self).__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch)
 
-    def compute_transformation(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.compute_dflip_transformation3d(input, params)
+    def generate_parameters(self, batch_shape: torch.Size) -> AugParamDict:
+        return AugParamDict(dict(params={}, flags={}))
 
-    def apply_transform(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.apply_dflip3d(input, params)
+    def compute_transformation(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.compute_dflip_transformation3d(input)
+
+    def apply_transform(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.apply_dflip3d(input)
 
 
 class RandomAffine3D(AugmentationBase3D):
@@ -242,7 +235,7 @@ class RandomAffine3D(AugmentationBase3D):
         resample: Union[str, int, Resample] = Resample.BILINEAR.name,
         return_transform: bool = False, same_on_batch: bool = False, align_corners: bool = False
     ) -> None:
-        super(RandomAffine3D, self).__init__(p, return_transform=return_transform, same_on_batch=same_on_batch)
+        super(RandomAffine3D, self).__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch)
         self.degrees = _tuple_range_reader(degrees, 3)
         self.shear: Optional[torch.Tensor] = None
         if shears is not None:
@@ -274,16 +267,16 @@ class RandomAffine3D(AugmentationBase3D):
         f"resample={self.resample.name}, align_corners={self.align_corners}"
         return self.__class__.__name__ + f"({repr}, {super().__repr__()})"
 
-    def generate_parameters(self, batch_shape: torch.Size) -> Dict[str, torch.Tensor]:
+    def generate_parameters(self, batch_shape: torch.Size) -> AugParamDict:
         return rg.random_affine_generator3d(
             batch_shape[0], batch_shape[-3], batch_shape[-2], batch_shape[-1], self.degrees,
             self.translate, self.scale, self.shear, self.resample, self.same_on_batch, self.align_corners)
 
-    def compute_transformation(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.compute_affine_transformation3d(input, params)
+    def compute_transformation(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.compute_affine_transformation3d(input, params['params'])
 
-    def apply_transform(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.apply_affine3d(input, params)
+    def apply_transform(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.apply_affine3d(input, params['params'], params['flags'])
 
 
 class RandomRotation3D(AugmentationBase3D):
@@ -338,7 +331,7 @@ class RandomRotation3D(AugmentationBase3D):
         interpolation: Union[str, int, Resample] = Resample.BILINEAR.name,
         return_transform: bool = False, same_on_batch: bool = False, align_corners: bool = False
     ) -> None:
-        super(RandomRotation3D, self).__init__(p, return_transform=return_transform, same_on_batch=same_on_batch)
+        super(RandomRotation3D, self).__init__(p=p, return_transform=return_transform, same_on_batch=same_on_batch)
         self.degrees = _tuple_range_reader(degrees, 3)
         self.interpolation: Resample = Resample.get(interpolation)
         self.align_corners = align_corners
@@ -347,12 +340,12 @@ class RandomRotation3D(AugmentationBase3D):
         repr = f"(degrees={self.degrees}, interpolation={self.interpolation.name}, align_corners={self.align_corners}"
         return self.__class__.__name__ + f"({repr}, {super().__repr__()})"
 
-    def generate_parameters(self, batch_shape: torch.Size) -> Dict[str, torch.Tensor]:
+    def generate_parameters(self, batch_shape: torch.Size) -> AugParamDict:
         return rg.random_rotation_generator3d(batch_shape[0], self.degrees, self.interpolation,
                                               self.same_on_batch, self.align_corners)
 
-    def compute_transformation(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.compute_rotate_tranformation3d(input, params)
+    def compute_transformation(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.compute_rotate_tranformation3d(input, params['params'])
 
-    def apply_transform(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return F.apply_rotation3d(input, params)
+    def apply_transform(self, input: torch.Tensor, params: AugParamDict) -> torch.Tensor:
+        return F.apply_rotation3d(input, params['params'], params['flags'])
