+++ kornia/augmentation/functional/functional3d.py
@@ -71,25 +71,14 @@ def apply_hflip3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch
     Args:
         input (torch.Tensor): Tensor to be transformed with shape :math:`(D, H, W)`, :math:`(C, D, H, W)`,
             :math:`(*, C, D, H, W)`.
-        params (Dict[str, torch.Tensor]):
-            - params['batch_prob']: A boolean tensor that indicating whether if to transform an image in a batch.
-                Example: With input batchsize of 4, only the first two tensors will be transformed if
-                batch_prob is [True, True, False, False].
 
     Returns:
         torch.Tensor: The horizontal flipped input
     """
-    # TODO: params validation
-
     input = _transform_input3d(input)
     _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
 
-    flipped: torch.Tensor = input.clone()
-
-    to_flip = params['batch_prob'].to(input.device)
-    flipped[to_flip] = torch.flip(input[to_flip], [-1])
-
-    return flipped
+    return torch.flip(input, [-1])
 
 
 def compute_hflip_transformation3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
@@ -108,62 +97,45 @@ def compute_hflip_transformation3d(input: torch.Tensor, params: Dict[str, torch.
     """
     input = _transform_input3d(input)
     _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
-    to_flip = params['batch_prob'].to(input.device)
     trans_mat: torch.Tensor = torch.eye(4, device=input.device, dtype=input.dtype).repeat(input.shape[0], 1, 1)
     w: int = input.shape[-1]
     flip_mat: torch.Tensor = torch.tensor([[-1, 0, 0, w],
                                            [0, 1, 0, 0],
                                            [0, 0, 1, 0],
                                            [0, 0, 0, 1]])
-    trans_mat[to_flip] = flip_mat.type_as(input)
 
-    return trans_mat
+    return flip_mat.repeat(input.size(0), 1, 1).type_as(input)
 
 
-def apply_vflip3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
+def apply_vflip3d(input: torch.Tensor) -> torch.Tensor:
     r"""Apply vertical flip on a 3D tensor volume or a batch of tensors volumes with given random parameters.
     Input should be a tensor of shape :math:`(D, H, W)`, :math:`(C, D, H, W)` or :math:`(*, C, D, H, W)`.
 
     Args:
         input (torch.Tensor): Tensor to be transformed with shape :math:`(D, H, W)`, :math:`(C, D, H, W)`,
             :math:`(*, C, D, H, W)`.
-        params (Dict[str, torch.Tensor]):
-            - params['batch_prob']: A boolean tensor that indicating whether if to transform an image in a batch.
-                Example: With input batchsize of 4, only the first two tensors will be transformed if
-                batch_prob is [True, True, False, False].
 
     Returns:
         torch.Tensor: The vertical flipped input
     """
-    # TODO: params validation
-
     input = _transform_input3d(input)
     _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
 
-    flipped: torch.Tensor = input.clone()
-    to_flip = params['batch_prob'].to(input.device)
-    flipped[to_flip] = torch.flip(input[to_flip], [-2])
-
-    return flipped
+    return torch.flip(input, [-2])
 
 
-def compute_vflip_transformation3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
+def compute_vflip_transformation3d(input: torch.Tensor) -> torch.Tensor:
     r"""Compute the applied transformation matrix :math: `(*, 4, 4)`.
 
     Args:
         input (torch.Tensor): Tensor to be transformed with shape :math:`(D, H, W)`, :math:`(C, D, H, W)`,
             :math:`(*, C, D, H, W)`.
-        params (Dict[str, torch.Tensor]):
-            - params['batch_prob']: A boolean tensor that indicating whether if to transform an image in a batch.
-                Example: With input batchsize of 4, only the first two tensors will be transformed if
-                batch_prob is [True, True, False, False].
 
     Returns:
         torch.Tensor: The applied transformation matrix :math: `(*, 4, 4)`
     """
     input = _transform_input3d(input)
     _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
-    to_flip = params['batch_prob'].to(input.device)
     trans_mat: torch.Tensor = torch.eye(4, device=input.device, dtype=input.dtype).repeat(input.shape[0], 1, 1)
 
     h: int = input.shape[-2]
@@ -172,55 +144,53 @@ def compute_vflip_transformation3d(input: torch.Tensor, params: Dict[str, torch.
                                            [0, 0, 1, 0],
                                            [0, 0, 0, 1]])
 
-    trans_mat[to_flip] = flip_mat.type_as(input)
-
-    return trans_mat
+    return flip_mat.repeat(input.size(0), 1, 1).type_as(input)
 
 
-def apply_dflip3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
+def apply_dflip3d(input: torch.Tensor) -> torch.Tensor:
     r"""Apply depthical flip on a 3D tensor volume or a batch of tensors volumes with given random parameters.
     Input should be a tensor of shape :math:`(D, H, W)`, :math:`(C, D, H, W)` or :math:`(*, C, D, H, W)`.
 
     Args:
         input (torch.Tensor): Tensor to be transformed with shape :math:`(D, H, W)`, :math:`(C, D, H, W)`,
             :math:`(*, C, D, H, W)`.
-        params (Dict[str, torch.Tensor]):
-            - params['batch_prob']: A boolean tensor that indicating whether if to transform an image in a batch.
-                Example: With input batchsize of 4, only the first two tensors will be transformed if
-                batch_prob is [True, True, False, False].
 
     Returns:
         torch.Tensor: The depthical flipped input.
     """
-    # TODO: params validation
-
     input = _transform_input3d(input)
     _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
 
-    flipped: torch.Tensor = input.clone()
-    to_flip = params['batch_prob'].to(input.device)
-    flipped[to_flip] = torch.flip(input[to_flip], [-3])
-
-    return flipped
+    return torch.flip(input, [-3])
 
 
-def compute_dflip_transformation3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
+def compute_intensity_transformation3d(input: torch.Tensor):
+    r"""Compute the applied transformation matrix :math: `(*, 4, 4)`.
+
+    Args:
+        input (torch.Tensor): Tensor to be transformed with shape (H, W), (C, H, W), (B, C, H, W).
+
+    Returns:
+        torch.Tensor: The applied transformation matrix :math: `(*, 4, 4)`. Returns identity transformations.
+    """
+    input = _transform_input3d(input)
+    _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
+    identity: torch.Tensor = torch.eye(4, device=input.device, dtype=input.dtype).repeat(input.shape[0], 1, 1)
+    return identity
+
+
+def compute_dflip_transformation3d(input: torch.Tensor) -> torch.Tensor:
     r"""Compute the applied transformation matrix :math: `(*, 4, 4)`.
 
     Args:
         input (torch.Tensor): Tensor to be transformed with shape :math:`(D, H, W)`, :math:`(C, D, H, W)`,
             :math:`(*, C, D, H, W)`.
-        params (Dict[str, torch.Tensor]):
-            - params['batch_prob']: A boolean tensor that indicating whether if to transform an image in a batch.
-                Example: With input batchsize of 4, only the first two tensors will be transformed if
-                batch_prob is [True, True, False, False].
 
     Returns:
         torch.Tensor: The applied transformation matrix :math: `(*, 4, 4)`
     """
     input = _transform_input3d(input)
     _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
-    to_flip = params['batch_prob'].to(input.device)
     trans_mat: torch.Tensor = torch.eye(4, device=input.device, dtype=input.dtype).repeat(input.shape[0], 1, 1)
 
     d: int = input.shape[-3]
@@ -229,12 +199,11 @@ def compute_dflip_transformation3d(input: torch.Tensor, params: Dict[str, torch.
                                            [0, 0, -1, d],
                                            [0, 0, 0, 1]])
 
-    trans_mat[to_flip] = flip_mat.type_as(input)
-
-    return trans_mat
+    return flip_mat.repeat(input.size(0), 1, 1).type_as(input)
 
 
-def apply_affine3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
+def apply_affine3d(input: torch.Tensor, params: Dict[str, torch.Tensor],
+                   flags: Dict[str, torch.Tensor]) -> torch.Tensor:
     r"""Random affine transformation of the image keeping center invariant.
 
     Args:
@@ -250,6 +219,7 @@ def apply_affine3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torc
             - params['syz']: Shear param toward y-z-axis.
             - params['szx']: Shear param toward z-x-axis.
             - params['szy']: Shear param toward z-y-axis.
+        flags (Dict[str, torch.Tensor]):
             - params['resample']: Integer tensor. NEAREST = 0, BILINEAR = 1.
             - params['align_corners']: Boolean tensor.
 
@@ -271,8 +241,8 @@ def apply_affine3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torc
     # concatenate transforms
     transform: torch.Tensor = compute_affine_transformation3d(input, params)
 
-    resample_name: str = Resample(params['resample'].item()).name.lower()
-    align_corners: bool = cast(bool, params['align_corners'].item())
+    resample_name: str = Resample(flags['resample'].item()).name.lower()
+    align_corners: bool = cast(bool, flags['align_corners'].item())
 
     out_data: torch.Tensor = warp_projective(x_data, transform[:, :3, :],
                                              (depth, height, width), resample_name,
@@ -296,8 +266,6 @@ def compute_affine_transformation3d(input: torch.Tensor, params: Dict[str, torch
             - params['syz']: Shear param toward y-z-axis.
             - params['szx']: Shear param toward z-x-axis.
             - params['szy']: Shear param toward z-y-axis.
-            - params['resample']: Integer tensor. NEAREST = 0, BILINEAR = 1.
-            - params['align_corners']: Boolean tensor.
 
     Returns:
         torch.Tensor: The applied transformation matrix :math: `(*, 4, 4)`
@@ -312,7 +280,8 @@ def compute_affine_transformation3d(input: torch.Tensor, params: Dict[str, torch
     return transform
 
 
-def apply_rotation3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
+def apply_rotation3d(input: torch.Tensor, params: Dict[str, torch.Tensor],
+                     flags: Dict[str, torch.Tensor]) -> torch.Tensor:
     r"""Rotate a tensor image or a batch of tensor images a random amount of degrees.
     Input should be a tensor of shape (C, H, W) or a batch of tensors :math:`(B, C, H, W)`.
 
@@ -320,6 +289,9 @@ def apply_rotation3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> to
         input (torch.Tensor): Tensor to be transformed with shape (H, W), (C, H, W), (B, C, H, W).
         params (Dict[str, torch.Tensor]):
             - params['degrees']: degree to be applied.
+        flags (Dict[str, torch.Tensor]):
+            - params['resample']: Integer tensor. NEAREST = 0, BILINEAR = 1.
+            - params['align_corners']: Boolean tensor.
 
     Returns:
         torch.Tensor: The cropped input
@@ -330,8 +302,8 @@ def apply_rotation3d(input: torch.Tensor, params: Dict[str, torch.Tensor]) -> to
     pitch: torch.Tensor = params["pitch"].type_as(input)
     roll: torch.Tensor = params["roll"].type_as(input)
 
-    resample_mode: str = Resample(params['interpolation'].item()).name.lower()
-    align_corners: bool = cast(bool, params['align_corners'].item())
+    resample_mode: str = Resample(flags['interpolation'].item()).name.lower()
+    align_corners: bool = cast(bool, flags['align_corners'].item())
 
     transformed: torch.Tensor = rotate3d(input, yaw, pitch, roll, mode=resample_mode, align_corners=align_corners)
